backend/pyproject.toml
[tool.poetry]
name = "djavue"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.4"
python-decouple = "^3.8"
psycopg2 = "^2.9.7"
django-extensions = "^3.2.3"
django-ninja = "^0.22.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.284"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"